pyproject.toml
[tool.poetry]
name = "hyena-pretrained"
version = "0.1.0"
description = ""
authors = ["pmaldonado4 <pmaldonado12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.24.3"
torch = "^2.4.0"
torchvision = "^0.19.0"
wandb = "^0.17.8"
hydra-core = "^1.3.2"
pytorch-lightning = "1.8.6"
ninja = "^1.11.1.1"
pandas = "2.1.0"
pyfaidx = "^0.8.1.2"
scikit-learn = "^1.6.1"
boto3 = "^1.36.23"
python-dotenv = "^1.0.1"
lion-pytorch = "^0.2.3"
torch-optim = "^0.0.4"

[tool.poetry.group.dev.dependencies]
transformers = "4.26.1"
omegaconf = "^2.3.0"
wandb = "^0.17.8"
matplotlib = "^3.9.2"
einops = "^0.8.0"
pyfaidx = "^0.8.1.2"
polars = "^1.6.0"
genomic-benchmarks = "^0.0.9"
loguru = "^0.7.2"
liftover = "1.3.1"
tqdm = "^4.66.5"
rich = "^13.8.0"
hydra-core = "^1.3.2"
opt-einsum = "^3.3.0"
cmake = "^3.30.2"
torchvision = "^0.19.0"
timm = "^1.0.9"
prettytable = "^3.11.0"
numerize = "^0.12"
git-lfs = "^1.6"
torchtext = "^0.18.0"
datasets = "^2.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
